Move submit button to navbar
